refactor(gateway): split handler.go into 4 files + optimize WriteCtx encode
Split handler.go (852 lines) by responsibility:
- handler.go: AEP dispatch + core routing (~335 lines)
- commands.go: control command handlers (~283 lines)
- worker_cmds.go: worker command routing (~248 lines)
- errors.go: error classification (~36 lines)

Performance: move aep.EncodeJSON outside c.mu.Lock in WriteCtx to
reduce lock hold time during control message writes.

Fix: elevate conn.go state transition failure log from Debug to Warn
as it signals potential state inconsistency.

Refs #352
